# add a clear_cache function!