--- a/bindings/python/src/models.rs
+++ b/bindings/python/src/models.rs
@@ -538,7 +538,7 @@ impl PyBPE {
     fn clear_cache(self_: PyRef<Self>){
         let super_ = self_.as_ref();
         let mut model = super_.model.write().unwrap();
-        if let ModelWrapper::Unigram(ref mut mo) = *model {
+        if let ModelWrapper::BPE(ref mut mo) = *model {
             mo.clear_cache()
         } else {
             unreachable!()