Fixes #1539 

Tried the provided script with this:
```python
In [4]: from transformers import AutoTokenizer
   ...: import gc
   ...: 
   ...: tokenizer = AutoTokenizer.from_pretrained("TinyLlama/TinyLlama-1.1B-Chat-v1.0", use_fast=True)
   ...: refresh_every = 100
   ...: 
   ...: for i in range(100000):
   ...:   s = f'{i} {i} ' * 10000
   ...:   tokenizer.encode(s)
   ...:   gc.collect()
   ...:   if i % 100 == 0:
   ...:     print(i)
   ...:   if i % refresh_every == 0:
   ...:     tokenizer._tokenizer.model.clear_cache()
```
and had no "leaks"