@ArthurZucker 2024-10-21
Yep, I can also do:
- give the ability to set the `cache_size`
- automatically set `cache_size` based on available RAM usage, which IMO would be better
-