@ecnuycxie 2024-10-24
The increase in memory usage has been bothering me for several days. looking forward to your fix for this issue!