@nixonjin 2025-02-21
I met an error:   AttributeError: 'tokenizers.models.WordPiece' object has no attribute 'clear_cache',  could someone tell me how to fix it?